online_service/requirements.txt
psycopg2-binary>=2.9.9
pyyaml>=6.0.1
Flask>=3.0.0
pyarrow>=17.0.0
s3fs>=2024.3.1
numpy==1.26.4
faiss-cpu>=1.8.0
pandas==2.1.4
gunicorn
